% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings-letter.R
\docType{methods}
\name{letter}
\alias{letter}
\alias{letter,ModString-method}
\alias{letter,ModStringViews-method}
\title{Subsetting a ModString to a single letter}
\usage{
\S4method{letter}{ModString}(x, i)

\S4method{letter}{ModStringViews}(x, i)
}
\arguments{
\item{x}{ModString}

\item{i}{the position of the letter to be returned}
}
\value{
a character vector
}
\description{
As the \code{\link[Biostrings:letter]{Biostrings}} functions, extracts a
substring from a \code{\link{ModString}} object by returning individual
letters by their position as a character vector.
}
\examples{
seq <- ModDNAString("AGCT/")
letter(seq, 5)
}
