% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings-ModString.R
\docType{class}
\name{ModString}
\alias{ModString}
\alias{ModString,MaskedModString-method}
\alias{ModString,AsIs-method}
\alias{ModString,ModString-method}
\alias{ModString,XString-method}
\alias{ModString,character-method}
\alias{ModString,factor-method}
\alias{as.character,ModString-method}
\alias{as.vector,ModString-method}
\alias{==,ModString,ModString-method}
\alias{==,ModString,XString-method}
\alias{==,XString,ModString-method}
\alias{ModDNAString-class}
\alias{ModRNAString-class}
\title{ModString objects}
\description{
The virtual \code{ModString} class derives from the \code{XString} virtual
class. Like its parent and its children, it is used for storing sequences of
characters. However, the \code{XString}/\code{BString} class requires single
byte characters as the letters of the input sequences. The \code{ModString}
extends the capability for multi-byte chracters by encoding these characters
into a single byte characters using a dictionary for internal conversion. It
also takes care of different encoding behavior of operating systems.

The \code{\link{ModDNAString}} and \code{\link{ModRNAString}} classes derive
from the \code{ModString} class and use the functionality to store nucleotide
sequences containing modified nucleotides. To describe modified RNA and DNA
nucleotides with a single letter, special characters are commonly used, eg.
from the greek alphabet, which are multi-byte characters.

The \code{ModString} class is virtual and it cannot be directly used to
create an object. Please have a look at \code{\link{ModDNAString}} and
\code{\link{ModRNAString}} for the specific alphabets of the individual
classes.
}
