% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings.R, R/datasets.R
\docType{data}
\name{Modstrings-internals}
\alias{Modstrings-internals}
\alias{seqtype,ModDNAString-method}
\alias{seqtype,ModRNAString-method}
\alias{seqtype<-,ModString-method}
\alias{seqtype<-,ModStringSet-method}
\alias{modsRNA}
\alias{modsDNA}
\alias{MOD_RNA_DICT_MODOMICS}
\alias{MOD_RNA_DICT_TRNADB}
\title{Modstrings internals}
\format{An object of class \code{DataFrame} with 162 rows and 9 columns.}
\usage{
modsRNA

modsDNA

MOD_RNA_DICT_MODOMICS

MOD_RNA_DICT_TRNADB
}
\description{
Analog to \code{Biostrings} there are a few functions, which should only
be used internally. Otherwise take care.
}
\keyword{datasets}
