% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings.R
\docType{package}
\name{Modstrings-package}
\alias{Modstrings-package}
\title{Modstrings: Working with modified nucleotide sequences}
\description{
Representing nucleotide modifications in a nucleotide sequence is usually done via special characters from a number of sources. This represents a challenge to work with in R and the Biostrings package. The Modstrings package implements this functionallity for RNA and DNA sequences containing modified nucleotides by translating the character internally in order to work with the infrastructure of the Biostrings package. For this the ModRNAString and ModDNAString classes and derivates and functions to construct and modify these objects despite the encoding issues are implemenented. In addition the conversion from sequences to list like location information (and the reverse operation) is implemented as well.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/FelixErnst/Modstrings/issues}
}

}
\author{
\strong{Maintainer}: Felix G.M. Ernst \email{felix.gm.ernst@outlook.com} (\href{https://orcid.org/0000-0001-5064-0928}{ORCID})

Other contributors:
\itemize{
  \item Denis L.J. Lafontaine \email{denis.lafontaine@ulb.ac.be} [contributor, funder]
}

}
\keyword{internal}
