% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings-ModStringSetList.R
\docType{class}
\name{ModStringSetList}
\alias{ModStringSetList}
\alias{ModDNAStringSetList}
\alias{ModRNAStringSetList}
\alias{ModDNAStringSetList-class}
\alias{ModRNAStringSetList-class}
\title{ModStringSetList}
\usage{
ModDNAStringSetList(..., use.names = TRUE)

ModRNAStringSetList(..., use.names = TRUE)
}
\arguments{
\item{...}{\code{\link{ModStringSet}} objects of one type.}

\item{use.names}{\code{TRUE}(default) or \code{FALSE}: Whether names of the 
input \code{ModStringSet} objects should be stored and used as the element
names in the \code{ModStringSetList}.}
}
\value{
a \code{ModStringSetList} object.
}
\description{
title
}
\examples{
mrseq <- c("ACGU7","ACGU7","ACGU7","ACGU7")
mrseq

# Example: contruction of ModStringSetlist from ModString objects
mr <- ModRNAString("ACGU7")
mr

mrs <- ModRNAStringSet(list(mr,mr,mr,mr))
mrs

mrsl <- ModRNAStringSetList(mrs,mrs)
mrsl

# Example: construction of ModStringSetlist from mixed sources
mrsl2 <- ModRNAStringSetList(mrs,mrseq)
mrsl2
}
