% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings-ModStringViews.R
\name{ModStringViews}
\alias{ModStringViews}
\alias{show,ModStringViews-method}
\alias{==,ModStringViews,ModStringViews-method}
\alias{==,ModStringViews,XString-method}
\alias{==,XStringViews,ModString-method}
\alias{ModStringSet,ModStringViews-method}
\alias{Views,ModString-method}
\title{The ModStringViews class extending the XStringViews class}
\usage{
\S4method{Views}{ModString}(subject, start = NULL, end = NULL, width = NULL, names = NULL)
}
\arguments{
\item{subject, start, end, width, names}{See 
\code{\link[Biostrings:XStringViews-class]{XStringViews}}.}
}
\value{
a \code{ModStringViews} object.
}
\description{
As the \code{\link[Biostrings:XStringViews-class]{XStringViews}} the 
\code{ModStringViews} is the basic container for storing a set of views on 
the same sequence (this time a \code{ModString} object).
}
\details{
For the details have a look at the 
\code{\link[Biostrings:XStringViews-class]{XStringViews}} class.
}
\examples{
seq <- ModDNAString("AGC6AGC6")
seq

v <- Views(seq, start = 3:1, end = 6:8)
v
}
