% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings-replaceAt.R
\docType{methods}
\name{replaceAt}
\alias{replaceAt}
\alias{extractAt}
\alias{extractAt,ModString-method}
\alias{extractAt,ModStringSet-method}
\alias{replaceAt,ModString-method}
\alias{replaceAt,ModStringSet-method}
\title{Extract/replace a substring from/in a ModString or ModStringSet}
\usage{
\S4method{extractAt}{ModString}(x, at)

\S4method{extractAt}{ModStringSet}(x, at)

\S4method{replaceAt}{ModString}(x, at, value = "")

\S4method{replaceAt}{ModStringSet}(x, at, value = "")
}
\arguments{
\item{x}{a \code{\link{ModString}} or a \code{\link{ModStringSet}} object.}

\item{at}{See \code{\link[Biostrings]{replaceAt}}.}

\item{value}{a \code{character} vector, \code{\link{ModString}} or a 
\code{\link{ModStringSet}} object. See \code{\link[Biostrings]{replaceAt}}.}
}
\value{
the modified \code{\link{ModString}} or \code{\link{ModStringSet}} 
object
}
\description{
\code{extractAt} extracts multiple subsequences from a 
\code{\link{ModString}} object x, or from the individual sequences of 
\code{\link{ModStringSet}} object x, at the ranges of positions specified 
thru at.

\code{replaceAt} performs multiple subsequence replacements 
(a.k.a. substitutions) in \code{\link{ModString}} object x, or in the
individual sequences of \code{\link{ModStringSet}} object x,
at the ranges of positions specified thru at.
}
\examples{
seq1 <- ModDNAString("AGCTTTT//TTTCGA")
seq2 <- replaceAt(seq1,IRanges::IRanges(7,10),"AGCT")
set1 <- ModDNAStringSet(c("AGCTTTT//TTTTCGA","AGCTTT//TTTTTCGA"))
set2 <- replaceAt(set1,IRanges::IRanges(7,10),"AGCT")
}
