% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings.R, R/Modstrings-ModString.R,
%   R/Modstrings-ModStringSet.R, R/datasets.R
\docType{methods}
\name{Modstrings-internals}
\alias{Modstrings-internals}
\alias{seqtype,ModDNAString-method}
\alias{seqtype,ModRNAString-method}
\alias{seqtype<-,ModString-method}
\alias{seqtype<-,ModStringSet-method}
\alias{XString,ModString-method}
\alias{XStringSet,ModStringSet-method}
\alias{modsRNA}
\alias{modsDNA}
\alias{MOD_RNA_DICT_MODOMICS}
\alias{MOD_RNA_DICT_TRNADB}
\title{Modstrings internals}
\format{An object of class \code{DataFrame} with 162 rows and 9 columns.}
\usage{
\S4method{XString}{ModString}(seqtype, x, start = NA, end = NA,
  width = NA)

\S4method{XStringSet}{ModStringSet}(seqtype, x, start = NA, end = NA,
  width = NA, use.names = TRUE)

modsRNA

modsDNA

MOD_RNA_DICT_MODOMICS

MOD_RNA_DICT_TRNADB
}
\arguments{
\item{seqtype, x, start, end, width, use.names}{used internally}
}
\description{
Analog to \code{Biostrings} there are a few functions, which should only
be used internally. Otherwise take care.
}
\keyword{datasets}
