% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modstrings.R, R/Modstrings-ModString.R,
%   R/Modstrings-ModStringSet.R, R/datasets.R
\docType{data}
\name{Modstrings-internals}
\alias{Modstrings-internals}
\alias{seqtype,ModDNAString-method}
\alias{seqtype,ModRNAString-method}
\alias{seqtype<-,ModString-method}
\alias{XString,ModString-method}
\alias{seqtype<-,ModStringSet-method}
\alias{XStringSet,ModStringSet-method}
\alias{modsRNA}
\alias{modsDNA}
\alias{MOD_RNA_DICT_MODOMICS}
\alias{MOD_RNA_DICT_TRNADB}
\title{Modstrings internals}
\format{
An object of class \code{DataFrame} with 162 rows and 9 columns.

An object of class \code{DataFrame} with 47 rows and 5 columns.

An object of class \code{DataFrame} with 170 rows and 3 columns.

An object of class \code{DataFrame} with 60 rows and 3 columns.
}
\usage{
\S4method{seqtype}{ModDNAString}(x)

\S4method{seqtype}{ModRNAString}(x)

\S4method{seqtype}{ModString}(x) <- value

\S4method{XString}{ModString}(seqtype, x, start = NA, end = NA, width = NA)

\S4method{seqtype}{ModStringSet}(x) <- value

\S4method{XStringSet}{ModStringSet}(seqtype, x, start = NA, end = NA, width = NA, use.names = TRUE)

data(modsRNA)

data(modsDNA)

data(MOD_RNA_DICT_MODOMICS)

data(MOD_RNA_DICT_TRNADB)
}
\arguments{
\item{seqtype, x, start, end, width, use.names, value}{used internally}
}
\value{
a XString* object
}
\description{
Analog to \code{Biostrings} there are a few functions, which should only
be used internally. Otherwise take care.
}
\keyword{datasets}
